改进SIESTA 单点计算模板, 方便用户替换default.fdf

+++ siesta/sp/submit.sh
@@ -7,13 +7,19 @@ source "$parent_dir/env.rc"
 # 1. prepare SIESTA input files
 # require gosh >= 0.0.25
 ln -s "$parent_dir"/psml/* . 2>/dev/null
-ln -s "$BBM_TPL_DIR"/default.fdf . 2>/dev/null
+
+# use default.fdf in BBM template dir, if it is not present in current job dir
+if [[ -f "$BBM_JOB_DIR/default.fdf" ]]; then
+    ln -s "$BBM_JOB_DIR"/default.fdf . 2>/dev/null
+else
+    ln -s "$BBM_TPL_DIR"/default.fdf . 2>/dev/null
+fi
 
 # Make SIESTA resume from last job. This could reduce electronic step costs.
 cp "$BBM_JOB_DIR/siesta.DM" . 2>/dev/null
 
 # 2. read fdf stream from stdin, write stdout to siesta.log for parsing
-tee input.fdf | mpirun siesta 2>&1 | tee siesta.log > "$BBM_JOB_DIR"/siesta.log
+tee input.fdf | run-siesta.sh 2>&1 | tee siesta.log > "$BBM_JOB_DIR"/siesta.log
 
 # 3. save important files to bbm job starting dir
 # require gosh >= 0.0.25
